@@ -1,43 +1,38 @@
 {
   "name": "stock-parser",
   "version": "1.0.0",
+  "type": "module",
   "engines": {
-    "node": "17.x",
-    "npm": "8.x"
+    "node": ">=18.0.0",
+    "npm": ">=8.0.0"
   },
   "description": "",
   "main": "app.js",
   "scripts": {
-    "start": "node server.js"
+    "start": "node src/server.js",
+    "dev": "nodemon src/server.js",
+    "vercel-build": "echo hello"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "api-error-handler": "^1.0.0",
-    "awilix": "^3.0.8",
+    "awilix": "^12.0.4",
     "body-parser": "^1.18.3",
-    "convert-excel-to-json": "^1.3.0",
-    "cors": "^2.8.4",
-    "create-react-app": "^1.2.1",
-    "eslint": "^8.14.0",
-    "express": "^4.16.3",
-    "express-fileupload": "^1.3.1",
-    "helmet": "^3.12.1",
-    "http-status": "^1.2.0",
-    "joi": "^13.4.0",
-    "mocha": "^10.0.0",
-    "mongodb": "^3.0.10",
-    "mongoose": "^5.1.3",
-    "morgan": "^1.9.0",
-    "node-rest-client": "^3.1.0",
-    "nodejs-dashboard": "^0.5.1",
-    "nodemon": "^2.0.16",
-    "request": "^2.87.0",
-    "request-promise": "^4.2.2",
-    "should": "^13.2.1",
-    "snazzy": "^7.1.1",
-    "spdy": "^3.4.7",
-    "supertest": "^3.1.0",
-    "uuid": "^3.2.1"
+    "convert-excel-to-json": "^1.7.0",
+    "cors": "^2.8.5",
+    "express": "^4.21.2",
+    "express-fileupload": "^1.5.1",
+    "helmet": "^8.0.0",
+    "http-status": "^2.1.0",
+    "joi": "^17.13.3",
+    "mongodb": "^6.13.0",
+    "mongoose": "^8.10.0",
+    "morgan": "^1.10.0",
+    "node-rest-client": "^3.1.1",
+    "uuid": "^11.0.5"
+  },
+  "devDependencies": {
+    "nodemon": "^3.1.9"
   }
 }
