@@ -2,11 +2,11 @@ export default {
   mongoUrl: 'mongodb+srv://balaji:balaji1234@cluster0-gqyfs.mongodb.net',
   mongoDb: 'stock',
   header: {
-    rows: 13
+    rows: 13  // Start reading from row 13
   },
   columnToKey: {
     A: 'quality',
     B: 'quantity',
     C: 'packets'
   }
-}
+}; 
